Add variable note lengths
Only reads in notes to score, not rests.
Instruments hard-coded to guitar and piano for now. Instrument selection coming
next.
Note selection is imperfect. If multiple notes are dragged over they get focus,
and the selection is not visible. This could be fixed by forcing blur or by
showing the selection and adding the ability to move notes.