Add users, sessions, login page, and register page
Add sessions to track whether a user is logged in or not and update the navbar appropriately.
Add users and credential handling for log in, sign up, and log out.
Add Login and Register pages with forms and server communication.
Add Api that initializes the application and handles credentials including storage and Http requests.
Update routing to show user profiles on `/username` and sessions on `/username/sessionName`.
Rename Session to MusicSession so that Session is available for user sessions.