Add Settings page
Add Settings page with forms to update password and email (part of issue #47).
Add Cognito methods to handle update requests and report success or failure.
Remove Username and replace with Account because we are storing a username and
email now.
Change Profile page to a placeholder. We will update account details on the
Settings page and to show public profile information.

There is one unresolved issue after this commit. When the Cognito method
updateUserAttributes is called, the token is not refreshed. We pull the email
to display from the token, so the user does not see their updated email.