Wire up profile to JavaScript API
Break Profile settings page into submodules. We will have  state management
in a couple of these which makes it worth breaking them apart.
Make a Profile type for profile information. Move avatar to it.
Start work in JavaScript Auth API to get the latest user values from Cognito
instead of off the token. This is implemented for the unfocused tabs and
DisplayName changes in this commit.
Make avatar url optional so that we can pass null from JavaScript.
Handle logouts in unfocused tabs when session changes.
Part of #50.