Update Auth API source of truth and add infobars
Update JS Auth API to call on Cognito for the latest user information. We can't
trust the token when a user updates their profile, so instead we get it from
Cognito.
Add infobars (aka toasts, banners) to display Auth API success and failure.
Infobars replace the functionality of AuthErrors.
Move onEnter form event to a FormHelpers module.
Minor improvements to error messages and styling.
Part of #50