Move infobars to Main and separate out navbar
Move the infobars to Main so that they can appear over the navbar.
In addition, this means they can persist even when we load in a new page.
Rework infobar to display top or bottom based on the current route.
Move the navbar to a separate module and change the way the logout route
is handled as a result.