Add public profile page
Add a public profile page. This will only work for the currently logged in
user until we add a backend route to get a profile.
Add validation that a user's website starts with an appropriate protocol.
Change the background color of the View Public Profile button.
Clean up unused imports.
Part of #50.